Safari-609.1.11.1/LayoutTests/fast/viewport/viewport-117-expected.txt

ALERT: viewport size 980x400 scale 0.88 with limits [0.88, 5] and userScalable true
